segunda parte .2

+++ escultor3d_SegundaParte/Putbox.cpp
@@ -1,6 +1,6 @@
 #include "Putbox.h"
 
-PutBox::PutBox(int x0,int x1,int y0,int y1, int z0 ,int z1,int r,int g,int b,int a)
+PutBox::PutBox(int x0,int x1,int y0,int y1, int z0 ,int z1,float r,float g,float b,float a)
 {
 xi = x0;
 xf = x1;
