+++ escultor3d_SegundaParte/Putbox.h
@@ -7,11 +7,11 @@ class PutBox:public FiguraGeometrica
 {
 protected:
     int xi,xf,yi,yf,zi,zf;
-    int r,g,b,a;
+    float r,g,b,a;
 
 
 public:
-    PutBox(int x0,int x1,int y0,int y1, int z0 ,int z1,int r,int g,int b,int a);
+    PutBox(int x0,int x1,int y0,int y1, int z0 ,int z1,float r,float g,float b,float a);
     void draw(Sculptor &t);
 };
